Added TinyTIFF lib, and fixed libQGLViewer
This commit adds the TinyTIFF image library for easily loading in
TIF/TIFF files, and doing it quickly. It can be changed at a later date,
although since it is entirely contained within two {h|c}pp files, it can
be compiled anywhere : on Windows, macOS, and Linux. As such, it makes
it very interesting for loading files in memory.