Finished basic TinyTIFF testing
In the future, we might need to do some more heavy-handed benchmarking
of the library, but for now we have the data we need.

The library is fast, efficient, and allows us to load TIF files at
anywhere from 0.8 to 0.05 milliseconds (!). It would allow us to load in
the entire stack of 4 thousand images in less than 2 seconds on average.