3D Texture loading (kinda) works (?)
It should be said, the first few images seem to load fine, but once we
get over the first once, boy does it degrade into a garbled mess.

Need to get over this problem. Shoudl'nt be the loading of images, got
this part down to a T (and slightly more optimized for execution time
than last week). Should look into moving to a RGB vector, not a
greyscale one.