Simple interactive viewer added
Added a simple interactive 3D texture viewer thanks to the work done on
the 3D_texture_test branch. The viewer allows to load and delete a
texture from a stack of images, and view it in real-time.

The loading of said textures might take some time if many images are
loaded (a thousand images usually takes around 15 seconds on my
machine).