Switch to GLM vectors, not qreal arrays
Simple switch from an array of qreal to represent positions and normals
to vectors from GLM, to get an easier time manipulating the vertices (to
possibly test out a feature to see the image stack in regular space as
opposed to transformed space as of now).

Some fixes in the CMake file to allow for the TIFF library to be
compiled only once, and linked afterwards (and to prevent it from being
recompiled using Qt's additionnal compilers [RCC, UIC, MOC]).