New image loader, new double viewer & fixes
This update brings a few additions and improvements :

- There is now a multi-threaded image loader. It is still a WIP, but it
is mostly done, I just have to iron out a few race conditions and
simultaneous access violations

- There is a new 'double viewer' underway. This allows for a scene to be
displayed from two different angles. The skeleton of the classes are
mostly there, I just have to add a scene class that will actually hold
the data to be displayed.

- TinyTIFF was updated to the latest commit (HEAD), not the one before
it (^HEAD) as it was before.

- A small library to handle OpenGL's VAO/VBO/shaders is coming in the
next few commits. It is greatly inspired by the GLSL helper classes in
KidPocketFramework, but updated to allow for a more modern use of OpenGL
shaders (with vertex array attributes & objects).