Switched TextureStorage for texture handling
The class ImageStorage was changed to TextureStorage, which holds within
it an image loader, to facilitate texture management at every level. It
allows to have as many TextureStorage(s) as we have stacks, and we will
be able to query a position within this stack with an interpolation
function. For now, the texel fetching is done with Nearest-Neighbor, and
the image loader is a bulk_texture_loader.

The Scene class was updated to reflect those changes.

The bulk_texture_loader class was updated to include a destructor,
freeing ressources upon deletion of the object.

The simple_shader_example test was removed from the project, as I have
figured out a couple of weeks ago what was causing the problem when
displaying things on screen using QOpenGLFunctions.