Added tetrahedral mesh (see prev commit)
In the previous commit, the architecture of the texture loader was
changed to allow for stack-by-stack allocation.

Along with this, this commit brings a tetrahedral mesh centered around a
point with which we can get a value by interpolating it with its
neighbors in real space, the values of which are taken from the initial
space.

Also, there is a WIP in the vertex shader to allow to see the
tetrahedras created.

The other files were just modified to accomodate the new tetrahedral
mesh, and to fix compilation and runtime errors.