Moving stuff between TetMesh and TextureStorage
Right before this commit, the organisation of TetMesh and TextureStorage
was kind of messy. Some code was executed one part in TetMesh, one part
in TextureStorage, some code was not executed at all (...)

This fixes it. Along with this commit and maybe the few following it,
the architecture now looks something like that :

- TetMesh, since it is defined in initial space, will only accept
coordinates defined in initial space. There is no transformation done
when providing a position to a TetMesh object.

- TextureStorage, since it represents a stack of images we have to
undistort to be able to reconstruct, will be defined as a grid of sizes
[width, height, depth], with two matrices attributed to it : one to
transform point from initial space to real space (which is the one
computed in the Scene::computeTransformationMatrix() function), and its
inverse, to transform points from real space to initial space. All of
the transforms will be done here, in this class since it is the one that
most interacts with the real space.