Voxel grid started, next is interpolation
A simple voxel grid implementation for computation has been added.
It is pretty barebones, but will eventually support rendering a
particular region, as well as writing to TIF files and DIM/IMA files.

Next commit will probably be the implementation of interpolation in the
tetrahedral mesh constructed in `TetMesh.hpp`.