Writers for the grid are being implemented
Multiple writers for the grid are being implemented. Right now, we have
:
  - GenericGridWriter : a simple, generic 'interface' class to write a
  voxel grid to a file on disk. Implements almost nothing, but exposes
  some handy methods for the classes ineriting it.
  - DIMWriter : writes the whole grid at once, in a DIM/IMA (Brainvisa)
  filetype. (Fully implemented, needs testing)
  - SingleTIFFWriter : writes the whole grid as a single TIFF file with
  multiple frames. (Under implementation).
  - StaggeredTIFFWriter : writes a voxel grid as a series of TIFF files,
  one for each depth level in the grid. Those TIFF files have only one
  frame, and the goal is to have the writing done for this on a separate
  thread, so that at each depth level computed, we immediately start
  writing the TIFF file associated.

Right now, each of those classes are 'fixed-type' classes. In the
future, the goal is to have those templated in order to be able to write
something other than arrays of unsigned char (DIM/IMA and TIFF files can
also support floating point, but that needs to be studied for usefulness
with the folks over at Tulane).