Added functionnalities, fixed bugs
Much of the morning was spent fixing bugs in my installation of drivers
on my laptop, but I was able to get some work done in the afternoon.

Added a couple of slider to determine the min and max values for the
data (min/max of where the data is significant).
Fixed not being able to see the generated voxel grid.

New bug : cannot generate grids correctly, to fix ! (Nearest Neighbor
works 90% of the time, for 80% of the grid, we should aim for 100% :) )

Underway : IMA loader, to load DIM/IMA files once generated (to check
the generation is well done)

Up next : cut plane visualisation, and instanciated voxel visualisation.