Added a few functionnalities
This commit introduces a cutting plane-based view of the voxel grid and
of the image stack, in both real and initial space.

This also fixes the problem with the DIM/IMA writer not writing the DIM
file (it was too small, under the buffer size which never triggered a
flush).

Next up, ficing the nearest neighbor and trilinear interpolation on the
voxel grid. After that, or while doing this, there will also be a voxel
grid-like visualisation based on instancing.

And at the end of this, the code from tests/neighbor_visu/ will be
merged in the project's root folder, and made the main executable of the
program. Only after that will the 'neighbor-search' branch be merged in
the 'develop' branch.