Volumic-like visualization, and improvements
This commit enables a volumic-like visualisation of the image stack in
real space and in initial space. Although be warned, this method is
*very* unoptimized (basically just instanciates width*height*depth
cubes, and draws the ones which have data inbetween the texture bounds).

Also, originally switched to mirrored repeat for the image to enhance the
grid generation in trilinear interpolation mode. But it messed up the
nearest neighbor generation ... was forced to remove it. Might think of
something later.

Next commit will probably be the one to put all the tests/neighbor_visu
files in the project's root folder, distributing the files where they
should be and tidying up the README.md file.