Add DiscreteGrid, a generalization of voxel grids
The class `DiscreteGrid` has been added to the project, and it is a
generalization of the concept of a voxel grid. It will be used
thourought the program, in order not only to have a single voxel grid
structure for input and output grids, but also to have more of a
disconnection between the program, and its visualisation component.

Next up, adding the DiscreteGrid component to TetMeshes, and
writers/readers.