Interpolation on the new DiscreteGrids
While there are some coordinate changes left to do, the grid generation
methods work. The only thing that doesn't is the coordinate change from
grid to world on the input grids.

Next will prob. be a bit more readers (implementation of the TIFF
version, from bulk_texture_loader). As well as bugfixing the grid
generation.