Touching up the final DiscreteGrid implementation
DiscreteGrid's implementation still has some bugs, although it should be
related to transformations only. The work done at this commit is mainly
attempts to find out the root cause of the bug.

Some candidates include :
  - matrix-point multiplication in GLM not producing the same results
  when done left-to-right or right-to-left (thus messing up the
  transformations between affine spaces)
  - the transformation matrix computed in `Scene` containing the full
  T(real, initial) transformation not being able to be inverted (those
  matrices are called non-singular IIRC ?)
  - some transformation being applied twice, or none at all through
  messed up function calls (exemple : toWorld(toWorld(point_gridspace))
  would yield incorrect results, since toWorld just multiplies the point
  by the T(initial, real) matrix without any additionnal checks)

Those are the prime candidates for the grid generation algorithm not
working correctly. Other than that, everything has been proven to work
in a previous commit in this branch.