DiscreteGrid is nearly done !
DiscreteGrid's implementation is nearly bug-free. The problem has been
pinned down to the implementation using bounding boxes for the detection
of content and/or boundaries of the render window.

The problem exists within the parts of the program under the
ENABLE_DATA_FITTING macro. It could be an erroneous voxel dimension
computation, or simply an error in the voxel positions within the
bounding box.

To investigate further tomorrow.

This commit also includes some slight improvements for the image reader,
although those are not documented here.