DiscreteGrid problems, ter.
It seems that TetMesh or DiscreteGrid fails to sample accurately the
input grids in a periodic fashion, regardless of the transform or the bounding box.

Need some more testing. For now, loads of print statements are issued,
better not to render anything too large (>5k elements) unless you want
your system to slow to a crawl.