DiscreteGrid done ! Code cleanup next
DiscreteGrid's implementation is done, and mostly tested (still have a
couple of tests to run). Some cleanup of the code is still incoming,
mostly to remove all the macros generated while testing the
implementation, and also to document the code properly. Some stray
functions (read unused) might also get removed if I have the time.

Some optimisation might be in order for the grid to be quickly generated
... Alas, it will have to wait.

Finally, it is done.