Added grid pool, grid list view, and grid details
This commit introduces the grid pool, a storage for all input/output
grids necessary for a program run. This data structure does not depend
on Qt, or a user-facing widget, and can thus be used later, once the
offline reconstruction will be implemented.

This commit also introduces the GridView class, a custom widget to be
used later in the main Qt-based UI, to list all loaded grids in a
window. This widget shows some basic information aboout a grid, and can
be used in coordination with a GridDetailedView to show and edit a
grid's properties.

This commit also removes the X/Y/Z controls at the bottom of the
neighbor-visu viewer, removes unnecessary functions in Scene, the
viewer, and a couple more classes.

Next commit will probably be some additions to the DiscreteGrid class,
generalization of the loaded content (in order to allow for offline grid
generation, or in-memory generation seemlessly, or even a (more)generic
approach to the transformation of a grid).

I'm getting back into a better workflow day by day. Got to keep it up :)