Removed tests and unnecessary files
This commit removes the tests done earlier in the year (all under the
tests/ folder). It also removes the header/source files which are not
used anymore.

'neighbor_visu' will become the main executable, and the next commit
will reflect that change, along with some improvements, of course. All
the unused targets have been removed from the CMakeLists.txt file.