Added grid detailed view, and bug fixes
This commit intriduces the GridDetailedView class, allowing to view the
details of a grid, and be able to read/modify them according to if the
grid is modifiable or not.

This new class interacts with the GridView class (bad name, might
change) in order to allow for a seamless experience when modifying a
grid's parameters (name only for those non-modifiable).

The files features.hpp and macros.hpp were also added, in order to give
a centralized control over what features are enabled/disabled while
testing the software.

The bounding box class has been restricted to only accept floating-point
numbers, thans to the use of the std::enable_if<> template. More details
in the macros.hpp and grid/include/bounding_box.hpp files.

Some more features were added and bug fixed, but honestly it's 6PM on a
Friday, those can wait.