Cleaned up Scene class, added cutting plane render
N.B. : The cutting plane render is partial for now, still need to debug
it on my desktop machine.

Added a 4-way visualisation to the main window, which for the moment
displays 4 POVs of the same scene. Will be changed later for 3 planar
cuts in the volume and one 3D/textured render.

Cleaned up the Scene class, removed useless attributes in the attribute
list in-shader and removed the approprpiate locations in the C++ program

Cleaned up the application's control panel, related to the changes in
the scene class.