The cutting planes are working ! (hard-coded)
For now, the cutting planes work, but for one grid only. Everything is
customizable for a given grid, but they are hard-coded from the C++
side. Should be able to work around it fairly easily (don't quote me on
that).

But for now, this will do. Work will resume on monday with the task of
showing the planes in 'constrained' view. The computation and
shader-isation is there, we have to consolidate it into a single
package.