Planar viewer done and tested
The PlanarViewer class has been implemented and tested. It allows to
visualize only one plane of a given texture (see TODO at the end). This
visualization of a single plane adapts to the framebuffer its being
displayed in, in order to show the entire texture at once, in
world-space, from one of the sides.

It takes the values available in the colorScale in order to display a
fragment's color, and responds to the changes in the mix/max value
displayed.

TODO : Allow to select another class, from the Scene class. Right now
hardcoded to the input grid.