First implementation of the texture3D method
This commit introduces a first try at the 'texture3D' visualization
method, showing the texture as a 'true' voxel grid, by essentially
raycasting into it using an acceleration structure (tetrahedral mesh in
this case).

For now, there's still some bugs. The tetrahedras are rendered, but for
some reason the raycasting part won't work. Will update as I go along.

This is a checkpoint where every buffer and texture is working, and the
only problems are (supposed to be) in the shaders.