Volum.visu (nearly) working, but needs cleaning up
Volumetric visu technically works, but had very poor performance. Should
see if ther's a quick way to improve it (cull unneeded tetrahedra from
CPU side ?).

Also, introducing a geometry shader for the volumetric visualization.
Sole purpose : display a wireframe overlay of the drawn tetrahedra at
relatively little cost.

I need to clean up the code, right now this is kind of a fixed-function
pipeline, where as we need to at least give the user the chance to load
what they want, when they want it.

Also, added the ability to only draw a grid's data on the cutting planes
when necessary.