Implemented the 'realistic' colour grading
This commit changes the loading pipeline a bit, in order to load 2 colour
channels of the same stack of images (in this case, a Red and a Blue
channel of the same tissue acquisition). It then correctly colours the
image stack in order to simulate a Eosine/Hematoxylin colouring of the
tissue sample.

[Colour equation taken from this paper : https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0159337]