Reshaping of UI/UX elements
New headers have been added above plane viewers, allowing the user to
directly control the position and direction of the cutting planes.

Some work has been done on the rotation of cutting planes, however, it
does not currently work as expected and will be fixed next week.

The automatic rescaling of the color scale will be added next week.
Currently, it scales automatically between two values, however those
values are user-defined (and user-modifiable at runtime). We should
analyse the 3D texture as it's being loaded, and get the min/max values
that way. The user will still have control of the visible domains,
however they will not control the min/max values of the color scale,
which are used to compute colors for each grid value.

There still persists the problem of the grid not being rendered properly
if some vertices are below the XY plane (i.e., Z < 0). Will have to take
a look at it next week.