Decoupled the color scale from original spinboxes
The color scale was originally normalized using the min/max texture
values the user chose. Now, the min/max of the color scale is determined
by another set of spinboxes, which propagate the color bounds to all
necessary shaders.

We still need to fix some issues with the planes (render only their
outer edge, of the right color) and with the bounding boxes.