Added ability to share context explicitly
The QOpenGLWidget documentation states that setting the attribute
Qt::AA_ShareOpenGLContexts to QCoreApplication is sufficient to make the
different viewers share the same underlying data, but it didn't seem
that way in practice.

Might be an error on my part, but it doesn't really matter now, once the
scene has begun to be/is initialized, the different viewers will have
their context re-created with the original context shared with
QOpenGLContext::setShareContext(QOpenGLContext*).