De-coupled Scene's init with image loading.
Still need to add the hooks to NeighborViewer to load on user input, but
we're getting there !

Simplified the texture upload process to OpenGL, with helper structs
(viewer/include/viewer_structs.hpp).

Added a QSurfaceFormat option to enable GL_KHR_debug if the extension is
present in the GPU driver (enables OpenGL logging). Still need to add a
QOpenGLDebugLogger with its init and startLogging functions to enable
APi validation.