Header guards and automatic mesh generation
The windows-specific variables in the main file have been surrounded by
header guards, to allow seemless compilation on Windows/Linux.

Also, added back an automatic mesh generation algorithm. It was
available on earlier versions, but removed for performance/dev reasons.
It has been re-enabled in this commit, but generates a mesh of fixed
resolution, no matter the underlying image dimensions.