Updated TinyTIFF, and CMake system along with it
TinyTIFF has undergone a few big changes. Its version 3.0 brings in
improvements to TinyTIFFWriter, in order to write multi-sample
multi-planar images.

It also changed its build system, relying on CMake. Unfortunately, it
does not support including it as a subproject, and instead needs to be
built separately. As such, we need to compile it beforehand ith
configure.sh (need to add configure.bat for windows users).

Also, updated Reader/Writer classes to reflect the slight signature
change of the loading/writing functions.