Readers now with shared_ptr, and grid reset on add
Whenever we add a new grid from the app, other ones are discarded.
This is done to prevent too much data from being loaded at once, and
also because we don't have use cases where more than one grid is needed.

Also, grid readers moved from raw pointers to shared_ptr, in order to be
added to the discretegrid class.
This is a hotfix to address the lack of proper offline grid generation
support. It will be completed in a few commits.