Modular (online/offline) loading/writing of files
This is a big commit which re-organizes the reader/writer classes in
order to allow for seamless loading/writing of files to and from the
program directly.

The user can set a memory limit which expresses the maximum size of the
images loaded in memory.

Also, the generation of grids now buffers one slice in memory before
writing it, allowing for a less memory-intensive generation.