Added new grid loader UI, modifying transformation
The user can now specify the parameters of the transformation when
loading the data, instead of it being hardcoded into the grid loading
mechanisms.

There's still stome bugs to squash, including a very weird one where the
OpenGL framebuffer will cause a crash when calling update() from the
scene control, but we'll get through it.

Also, moved the as_const<>() template to macros.hpp, in case we need it
elsewhere (needed it for a few changes in loader_widget.cpp).