Bugfixes for 16-bit images
This commit squashes the program-breaking bugs, although some more
subtle ones might remain.

Loading of 16-bit images is now nearly fully possible, and seamless from
the original 8-bit variants.

The bug was that I was reading past-the-end in visibilityMap in Scene,
since I was iterating from [0,  65535] in an array defined in [0, 255]