Add libtiff as a dependency
Also, added a CMake file to find the local version of the library, and
link it against the targets that need it.

Modified the configure.sh script to take into account the libTIFF
compilation that must also take place before the main project is
compiled.

There's still the matter of trying it on windows, after updating the
configure.ps1 script.