Added ROI user input and ROI rescaling at runtime
This commit introduces the possibility for users to define a 'ROI' (a
subset of available domain values) at loading time.

This 'ROI' can then be changed at runtime, and will the changes will be
reflected not only on the computed color scales, but also on the sliders
at the bottom of the program.