Renamed preComputeImageData() to parseImageInfo()
The renaming was done purely to be more precise about the action done in
each function overload. The real 'precomputation' is actually done in
preAllocateStorage(), where the image dimensions and voxel sizes are
adjusted to the level of downsampling chosen.

Also, contains some of the first few changes to implement the libTIFF
reader :
- some modifications of the libTIFFReader::TIFFFrame::loadTIFInfo()
  function, to parse the planar configuration first (tiled or stiped),
  then the sample format, then the other data relevant to the image.
- a first draft of a read cache was implemented. It's a templated class
  taking an index type and a data type, to be a bit more flexible.
  Nothing has been implemented as of yet for this cache structure.