Multi-threaded loading, v1
This commit enables a simple multithreaded loading of images, presenting
a direct feedback of the loading process to the user in the form of a
loading bar.

This launches one thread if loading one stack of data, an dtwo if
loading two.

This might be improved later to take advantage of more cores.