bug fix for offline generation
The offline generation process was not only not reading enough data, but
it was also not setting the right directory for the tiff files.

fixed by implementing those features and fixing up the generation
process. next, we need to make it run in the background, like the grid
loading part.