Multi-threaded parsing of files added
The parsing of the files (before opening them) is now multi-threaded
[kinda].

Like with the loading itself, it happens on another thread than the main
GUI one, and allows to get a more responsive program.

When loading large files or huge amounts of small files from a slow
storage media like a hard drive, the program would freeze and make the
user wait needlessly.