OME-TIFF testing & UI elements
The OME-TIFF parsing method has been tested a bit more thouroughly, and
seems to be working as intended.

Some UI elements have been added in order to load the OME-TIFF files (in
comparison to regular TIFF files). Two buttons have been added to the
loader widget, which allow the user to specify which kind of TIFF reader
to use for the files.

If the user selects the (regular) TIFF reader, but then loads a OME-TIFF
file, the file will not be parsed according to the OME-TIFF
specification, relying instead on basic TIFF parsing methods.

In addition, the voxel dimensions will automatically be updated when the
file format specifies it (in the case of OME-TIFF, for example).