View to coordinates, and bugfixes
This commit introduces a few features :
  - the ability to get a coordinate from any of the viewpoints (3D,
    planar, volumetric ...) in world space
  - the ability to _properly_ load a ROI of your grid in the program
    (there were some bugs last time, they are fixed now)

Some of those features were started in previous commits, but were only
concluded now. All that's left to do for the coordinate fetching is
implement some way to show in 3D (or maybe just in the planar viewers ?)
where the point selected is.

It should not be too hard to implement a zone select. To determine how
useful that would be right now, or if we can wait a bit before doing it.