Adding simple image cache
A (very simple) image cache has been added to the reader class. This
cache is for the moment only present on the libTIFF implementation of
the grid reader (and by extension, the OME-TIFF reader as well).

The rationale behind it is because the DIM/IMA reader need not open and
close files at will everytime we read a pixel (since the arrangment of
data in the file is known ahead of time), there's no need for a cache.

Next commit will introduce a ANALYZE/NifTI reader backend.