Added NIFTI lib, updated config and CMakeLists
The official NIFTI library has been added in the project.

As the root folder of this project was starting to get crowded, all
third party libraries have been moved to a folder of their own, with an
updated configuration script for linux (windows coming soon).

As a bonus, the config script automatically compiled _all_ libraries
which need to be included, and installs them locally in a
compiled_libraries/ folder which can be externally linked (simplifying
the cmake files to find local libraries). Also simplifies the
compilation process, since QGLViewer is automatically compiled and
installed locally.