Update for Windows compilation
The Windows version of the configuration script has been updated to reflect the new architecture of submodules within the project.

libTiff and TinyTIFF compile without issues, and the other two have problems. NIFTI cannot find its own headers (Znzlib, included right in their files is not found). Meanwhile, QGLViewer straight up does not compile the same on Windows. The `make install` procedure only copies the designer plugin to the installation directory, and that's if it even managed to compile.

It's true that compiling using QtCreator is a breeze, however it only compiles the library locally. As such, the CMake file has been updated to either sample the files from the `third_party/compiled_libraries` folder if on Linux, and from `third_party/libQGLViewer/QGLViewer` if on Windows.

The script will be updated in the future, with more up-to-date code if the need arises.