Starting to implement the grid re-worked interface
This commit introduces a first version of the grid backend interface
that we will use for all other derived implementations. It also
introduces a stub for the grid API, and a slightly re-organized file
structure, to move away from project-wide dependencies in the CMake
lists, and towards a more fine-grained dependency graph with the
introduction of the new grid API.

The class ThreadedTask has been shamelessly stolen from the IO
namespace, in the file reader.hpp, since it provided a nice enough way
to interact with a process on another thread.

And finally, the implementation of the TIFF backend has begun.