Further work on the grid and image backends impl.
More functions have been added for the backend, and the grid interface
is coming along nicely.

The TIFF backend has seen the most work today, and we'll have to
re-think how to handle OME-TIFF files. Currently (in the non-reworked
program) they are a class which inherits the libtiff reader, purely to
implement a custom parsing class.

But the TIFF classes need to be subclassed with templated readers, so
one alternative would be to implement the read logic of TIFF files as a
pImpl class, with the templated class as the pImpl.

It's pImpl all the way down, baby.