Second pass over the implementation - begin tests
This is the result of a second look at the implementation of the TIFF
backend. The testing of the class can now begin.

Furthermore, a README file has been added to `<root>/image/tiff/`,
detailing the kind of data and TIFF files supported by the
implementation. This will be updated if the requirements/supported
formats change.