Testing on the grid revealed a flaw
Due to a misplaced used of `auto` in one of the examples i made before
starting this endeavour, the implementation needs to change slightly.

Notably, when defining a template member function in an abstract class,
this function _will not_ get overloaded with a function of the same name
in a derived class. Damn.

As such, when calling the function `foo<T>()` on an object of type
`std::shared_ptr<Base>` which contains a pointer to `Derived`, the
function called will be `Base::foo<T>()` (which is non-defined) instead
of the "overloaded" function `Derived::foo<T>()`.