Added axes drawn at the position of the click
Added a simple set of axes drawn where the user clicks to get the
position in image space.

Additionally, those axes can either be drawn like any other mesh/image
in the viewer, or be draw on top of whatever's on screen.

Also, added a message in the statusbar to tell the user the image
coordinates of the point under its mouse.