Fixes for OpenGL 3.2 & GLSL 1.50
In order to be compatible with more hardware and software configurations, we want compliance with OpenGL 3.2 and GLSL 1.5. The only "exotic" extensions still required are GL_ARB_explicit_attrib_location which is _required_ for the shaders to work. This is the extension that made the `location() <value>` syntax available to shader programmers.