Added simple transforms to the new grid API
One main feature the new grid API lacked over the old one was the
support for arbitrary transforms. This commit introduces a basic
transform class, and the concept of a transform stack.

Those concepts will later be expanded upon to include more transform
types, but for now those will suffice.

Also, slightly changed the scene init, thought it would change the error
messages in renderdoc (improper VAO use, not generated properly). It did
not, in fact.