Added/tested transforms, fixed bug in ThreadedTask
ThreadedTask would sometimes deadlock, return a "not permitted" error,
or crash the program outright. This was due to a wrong use of mutexes
using the std::mutex API. The locks were scoped with the object, meaning
they could be accessed by multiple threads, leading to potential
deadlocks.
Fixed it by making each call to a task function having its own lock.
Since the same thread cannot call concurrently two different methods,
and since every method releases the lock once finished, this ensures no
deadlock can occur from my end.

Also, the parsing function did not correctly report the progress of the
parsing of the images. The TIFFBackend::checkFilenamesAreValid()
function would return the number of frames for the grid (as if one
channel were present), but the parsing function incremented the current
progress on every frame of every color channel/image stack. This lead to
bugs in the LoaderWidget class, since the queries to the grid were
effectively done before it was parsed. This is now fixed.

Tested the grid transform stack integration into the program. It works
as expected to represent transformations, however the visualization of
the dataset afterwards does not work. The volumetric mode is stretched
(as if the voxel dimensions were not respected) and the solid mode lacks
the visualization of the cutting planes. The 2D viewers also lack in
this department.