Working on allowing more TIFF files
Right now, we're pretty drastic : we only accept files with one sample
per pixel, and the only way to add more information channels is to add
more single-sample TIFF stacks to the original filenames. As such, it is
not really that user friendl. Sometimes, we may need to load RGB data
(as is the case with IGF and Tulane, both revealed around the same
time).

This commit introduces a bit of the work required to do so, but it's
going to take a bit of time.