Further implementation of TIFF rework
Need to test it, but has the same implementation as the original
structure. Should be able to just plug it straight into the code and run
with it.

Should probably also add a way for the user to specify data for certain
slices (in a separate, non-erasable cache). Pulls the original image(s)
and applies the user-defined data on top.

After that, the writer classes check if the slice has been modified by
the user. If so, get the user-defined version and if not, get the
on-disk versions.