Bugfixes for TIFF implementation & scene
Some library handles were not properly released, and the TIFF standard
apprently has a tag named 'SMinSampleValues' which _can_ be set
sometimes. Just in case, it is now set.

Also, fixed a bug where a non-valid grid would still be parsed when
creating it in the loader widget.

Also, added the ability to hide grids by unselecting a channel. That was
not done since the RGBMode enum in Scene was ignored with the UBO data.