Doc update & TIFF writer
Update the doc on the Scene class, for the few functions that weren't
already documented.

Finished the first implementation of the TIFF writer. Should probably
remove the constructors without a base path, but it's fine for testing
purposes.
Started a Scene::draft_saveFirstGrid() function to test the new writer
(dirty hack that I should remove soon).

Will continue to test it if I have the time.