Fixes for Windows compilation & OpenGL Core
The TransformType::Unknown enum was colliding with Windows' enums (more
specifically, the _MEDIA_TYPE enum).

Also, the <tiff.h> header could not be found, since it's in the standard
include locations on Linux, but on Windows we have to fetch the one from
the third party libraries.

Also, finally figured out how to force Qt to use a Core OpenGL profile.
When setting the default QSurfaceFormat (neighbor_visu_main.cpp), you
not only need to set the profile, but also the minor/major format.
Otherwise, it just defaults to OpenGL 2.x which has no notion of Core or
compatibility.