Add patch for Windows compilation of NiftiClib
---

Required because I cannot get for the life of me this library to compile
under Windows when Zlib is not installed system-wide. They look for the
ZLib package but do not include the ZLib headers in their CMake file of
the included znzlib. It's to the point where I wonder if they even
tested the library on Windows.