First draft of the generic_image_downsampler
---

This is a WIP for the generic image downsampler. It takes a pixel type
AND a functor in template arguments in order to determine the pixel data
types and the downsampling strategy respectively.

Ideally, it would allow us to perform an instanciation such as this :
  GenericImageDownsamplerTemplated<std::uint32_t, nearest_neighbor>()

This is still a very early development phase.