Downsampling & SubRegion : ctors and impls
---

Added createBackend() functions for sub-region and downsampled grids. It
allows for the ctors of the Grid class to actually have a valid pointer
to implementation with a GenericImageReaderSubRegion or a
GenericImageDownsampler templated instance attached to it.

Also, fixed a bunch of include paths that Thomas left untouched in the
great renaming effort done earlier.