Further work on the downsampling grid
---

This commit introduces a local cache for slices (in image/utils/include)
which holds as many slices as resquested. This allows for the grid to be
downsampled entirely at loading time, on the
GenericGridReader::parseImageInfo() function (and its overloaded
siblings).

This also introduces the generic_image_interpolators.hpp file, where
some implementations of default interpolators will be placed later.

The signature of interpolation functions has been changed to include 2
grids and an index : source/target grid and position in grid coordinates