Stub for default interpolator & fixed signatures
---

Added a default 'null' interpolator (makes everything 0) which is
selected if no valid enum value is given to the
GenericImageDownsampler::createBackend() function.

Also, added the Image::Downsampler::findRightInterpolatorType<>()
function which returns a suitable instance of interpolator functor for a
given data type and a given resampling technique (from enum values).

Added voxel sizes to the downsampled templated reader in order to not
perform the computation of the voxel sizes at each time the interpolator
functor is called (since it needs the voxel sizes of the target image
reader).