Scene control can now handle every data type and is updated at load
WARNING: for now sliders can handle only datatype lower or equal to
integers as their internal datatype are integers. Thus if the datatype in the image
is bigger than integer the possible values available in the sliders are clamps.

TODO: make sliders internal representation a ratio, allowing to use them with every datatypes
even if the internal data of a slide is integer.