Added shader compiler, and Mesh I/O
---

Utility class 'ShaderCompiler' was added, in order to provide a quick
and easy way to have shader compilation/linking happen, with the ability
to perform #pragma replacements.

Thus, we can replace the contents of a line containing '#pragma ...'
with the contents of a file, or a string of text. Tested in the Scene
class in order to perform shader compilation. Seems to work alright.